feat(finish): adding support for victory on reaching the finish
 Guardian must commit to the victory condition types, and various other types
including the choice types when creating the game. The transcript stack
proofs can then only work in terms of those types. And enough semantics
should then are be available to create a meaningful challenge and prevent
arbitrary behaviour by the dungeon guardian.

 Support for two transition types is now available. The first is the location
transition used to verify movement around the map. The second is the finish
transition. Furniture placement allows a LocationChoice to encode an input that
selects the finish.

 When the finish transition is resolved it will (soon) award ownership of the
transcript to the victor and complate the game. The victor will then also get
to mint nfts representing the remaining participants being "trapped" in the
chain.