feat(victory-transfer): transfer the game nft to the victor
The first participant to reach the placed finish causes the came to
complete. The ownership of the game session nft is transfered from the
guardian to the victory.

As currently implemented, the guardian can chose to never complate the
game. It may be good enough to rely on social presure here, but we also
intend to build in disinsentives